requirements.txt
langchain
langchain-community
transformers
torch
sentence-transformers
faiss-cpu
youtube-search-python
PyPDF2
streamlit
